New folder structure Divided components to smaller files and added folder hierarchy
