+++ src/assets/style/Footer.css
@@ -1,0 +1,1 @@
+/* TODO: ADD STYLE FOR FOOTER */
