+++ src/assets/style/index.css
@@ -1,0 +1,16 @@
+.user-data {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  font-size: clamp(0.5rem, 5vh, 2rem);
+  text-align: center;
+}
+
+/* TODO: ADD STYLE FOR INPUT AND BUTTON*/
+input {
+  width: 5vh;
+}
+
+/* button{
+
+} */
