+++ src/parts/Footer/Footer.jsx
@@ -1,0 +1,1 @@
+//TODO: ADD FOOTER WITH GITHUB LINK
